refactor: use composition to add behavior to tiles
Instead of having one big God tile, tiles now have a "Terrain" property.
This Terrain is a node with a specific script "WithXYZ" (for instance,
WithIsland).

A Terrain is generated whenever the Tile type changes from the editor,
or if the scene enters the tree and there is no Terrain/the terrain
does not match the type.