fix: resolve an issue where player could do an invalid move
When navigating to a WithCurrent tile, the player is supposed to be moved to the next tile
automatically by the current. But! If the player inputs another move too quickly, it was
possible to do an invalid move from the WithCurrent tile.

To prevent this, the input handler will now check if the tile can undock from the
desired direction. Sadly, I was unable to prevent the player from spamming inputs:
it's still possible to force an input while the WithCurrent tile is navigating on its own,
but at least this player input won't be in the wrong direction.

This is enough for now, but more refactoring would be great in the futre.