fix: improve the navigation path visual when moving through a WithCurrent tile
Previously, moving between tiles would fade the previous path. The implication is that
when moving from Island 1 to Island 2 through Current 1, the path would seems split between
Island 1 in one side and Current 1/Island 2 on the other side. To prevent this, we will now
expand upon the previous path when going trough a tile without player control (.ie, like
a WithCurrent tile).

An unexpected consequence of this change was that it was no longer possible to check for (and free)
a preexisting path with `SequenceEqual` when going back to a previously visited WithCurrent
because when the check is done at each step of the navigation (ie., when selecting Current 1,
then again when selecting Island 2), and in the first step, the new path isn't equal yet to the
old one until Island 2 is reached.
Thanksfully, using `Contains` instead does the trick, at least until another unexpected consequence
appears!