megafeat: this game is now about pirates... SOFTWARE PIRATES!
This is a big change. I left the pirate theme behind and replaced it with
a hacking theme instead. Two reasons for this:

1. It will be easier to do the visuals with this theme.
2. Tbh, I like the gameplay of the puzzles, but I wasn't happy
with the pirate integration: why would a pirate ship draw a
cross on each island on a map? Shoudn't it be just ONE island?
I didn't want to change the puzzles core concept, but it was
important to me to have a strong theme and fantasy to justify it.

From now on, the game is about hacking. Instead of marking
random treasures on island, the puzzle needs you to enable
every core in each level. The wrecks tiles are now firewalls,
the water currents are routers, and there is no longer a
boat (that's what I'll miss the most).

This commit includes big changes in assets and folders structure,
but I haven't started on the UI, and I have on tile left to work
on (the router), but most of the visuals are now less about
"yarr!" pirates and more about "l33t" pirates.

Happy hacking!