refactor: improve the querying of next selectable tiles
This big refactor was necessary to truly get the next selectable tile
in edge-cases when there is a redirection in the path (for instance,
when moving through a Router or a Proxy).
For instance, at line 165 of the method "GetNextSelectableTileInDirection",
we overload the current direction with the Router direction.
At line 166, we're replacing the adjacent tile by the proxy exit tile.

Tbh I'm not entirely satisfied with the pattern I used on this file; ideally,
instead of big switches in the Tile class, I want each specific terrain
to handle the logic on their own. I'll probably refactor this part again.