fix: prevent unwanted movement when on Router tile
So, this one is tricky: when moving to a Router, the game will immediatly
trigger an input in the Router's direction, and we'll move in a split second.
But it seems that if you're quick enough, you can move while on the Router, before
the automatic movement!
This isn't an issue on its own, but since the game assumed that the only moves
from a Router tile were allowed moves (ie. the input triggered programmatically),
there was NO CHECK to prevent illegal moves.
This should not be corrected.