refactor: use abstract class TileTerrain
It's about time! I talked about my wish to refactor
this part of the code (for instance in 2eb6334ecfed825a93f4c3d238379d240813e2661~)
and decouple the the logic of Tile terrains from the
Tile. This big change introduce an abstract TileTerrain
class, and let each Tile terrain maintain its own logic
instead of using a switch statement in the Tile.
Hopefuly, this should make it easier to add new Tile terrains
in the future!

nb: the "terrain" name made more sense when this was a
pirate sea, and I will probaby change it once I find
some thing better.