gameplay: add "Glitch" property to Core
When selecting a "glitched" CoreTile, the first CoreTile
in each direction (Up/Down/Left/Right) will be automatically toggled.
The neat thing is that the trigger use the same logic as the player:
this mean that a Router or a Proxy near the glitch will be used to
get the next core tile and toggled it!