visuals: BIG changes to LevelUI and MainMenu
This commit introduce new custom labels and buttons
under the "Kami" prefix, which will allow me to
compose themes easily by mixing sizes and colors.
To get to that point, I had to refactor a lot of code.

While I was on this part, I removed the zoomUI
from the DynamicCamera and put it into the LevelUI
instead.

Also, I added a hidden scale coefficient for the UI
on mobile (android/iOs) phones, since it was way
too small.

Yes, this commit is a mess: sadly, I spent two
weeks on it, hence why it's so disorganised.
Sorry, myself!