gameplay: prevent multiple inputs on touchscreen
It was too easy to make a mistake by dragging
more than wanted, so I choose to block input
once the first move is done and until the
dragEvent is over.

In the future, it would be cool if the player
could choose which way they want in the settings.