visuals: merge LevelUI and GridUI, rotate grid on mobile
1. By merging the LevelUI and GridUI, the layout will be
way easier to position. The grid is centered on the screen,
while the HUD sits directly on top of it. And the grid size
doesn't break this layout!

2. Since I want the mobile device to show a vertical grid
instead of a horizontal one, I had two options: either make
two versions of each level (!), one for desktop and one for mobile,
or wrote a function to rotate the grid by 90°. I choose the latter
to make my life easier moving forward.
But it makes more sense to have a vertical grid by default and
rotate it on desktop, since mobile devices are less powerfull.
I'll probably do this instead.