refactor: add Init virtual method to TileTerrain
This will prevent weird interactions when a Tile is duplicated
during the board rotation. For instance, a core with `isCoreEnabledOnStart`
set to true would toggle twice, first when the original Tile is ready
then once again when the duplicated and rotated Tile is ready.

From now on, TileTerrain setup will be moved to a similar Init method
to prevent this.