refactor: rotate Grid on Desktop instead of Mobile
As I said in 38db840ee8c81b5144532e14fa472384b66743fe1~,
I want to the default Grid orientation to be paysage instead
of portrait. This way, the game on mobile devices doesn't have
to do more work to setup a level. From now on, the rotation will
be done on desktop.

This means that I have to go back to every levels and remake them
in paysage, but this isn't so hard: I just have to run the level
once to rotate the original paysage to portrait, then copy the
scene at runtime and replace the original directly in the editor.