fix: resolve an issue with Grid Rotation
Previously, the Grid wasn't so much rotated as it was
shuffled around, ie. the columns and rows were swapped,
sure, BUT the tiles were put in the wrong order, which wasn't
supposed to happen.

Also, I added a Init method in the Router to check if the board
has rotated and if yes, to rotate the Router's Direction as well.