refactory: detect Proxy ExitTile automagically
Here's the issue: when rotating the board for desktop devices,
the Proxy tiles cant keep their ExitTile because its reference
is removed when the original tiles are freed. Not good!
Instead, we now have to set the same "ProxyId" value to two
proxies to linked them togehter, and the Init method will
use these to find the right ExitTile.

As a nice bonus, maybe this new ProxyId could be use to generate
a color variant on the Proxy portal? Will keep looking into it
later.