fix: resolve wrong StartingTile after board rotation
Funny thing: on some levels, the rotation was breaking
some tiles names. It's not a big deal most of the time,
but when the affected tile is supposed to be the starting
tile, this would break the level!

The bug was caused because tiles from different rows
may share the same name, for instance "Row1/Tile_0_0"
and "Row2/Tile_0_0". But after rotation, these two
tiles could be in the same row, and the engine
cant have two nodes with the same name and the same
parent, and thus the second tile was renamed something
like "@NodeType@"... and we would lose the real
starting tile. Oops!

I fixed this by checking for the starting tile before the
actual rotation. And while I was at it, I added the id
of the tile to its name, to avoid confusion between
rows.