levels: use new GridContainer on every level
This big commit adds the GridContainer to every level,
and rework the tiles to be in portrait mode by default
instead of landscape. This way, the mobile devices
don't have to do extra computations to rotate the grid,
leaving this burden to the desktop.

While I was on this, I did re-balanced some levels which were
either too big or didn't have the correct optimal score.