feat: open MainMenu seamlessly from a level
You can now open the main menu, then resume your current game.
While I was at it, I removed the (obsolete) LevelTitle and LevelSubtitle
and replaced same with an Id property, which will be use later
which will be used later to know the current level while we're
on the MainMenu and a game is playing in the background.