fix: setup the FirewallTracking arrows properly on startup
This big commit fixes a few issues with the FirewallTracking:
because of the asynchronous nature of the Level startup,
and since the Firewall (with tracking) is the only Terrain to
do logic based on the rest of the Level, I had to await
the end of Level startup, which was previously done in the
background while the Game started early. Oops!

This commit also introduces a new texture for the base Firewall.